MAT-131
Add an "--output" launching option for the cli and prometheus
that allows to decide where the logs will be created.
The default log file repertory is now "/tmp/logs".
If the output directory doesn't allow to write files, then the person who have decided
of this location should have checked beforehand and deserve to have a lot of logs.