MAT-357
High Power Consumption on hci-esx-01

* On the target, the sum of sub monitors' energy values leaded to
decreased target total energy due to missing monitors. Now the energy
parameter is always computed based on the sub monitors' power
consumption, and even if we encounter less power consumption because of
the missing monitors we are sure that the energy is going to grow up as
the  `CollectHelper.collectEnergyUsageFromPower` will take care of the
energy incremental sum.