MAT-364
EMCNavisphereCLI : discovery and collect do not work

Not using process.waitFor(timeout, TimeUnit.SECONDS) any more because it never return with some commands.
Have to poll it in reading the inpustram regurarely before using it but in this case the timeout is not right.
So better using a Future with timeout with a simple waitFor at the end.