HWS-459
Addition of "Last Resort" Connectors

New hdf directive specifying that a connector can be used as “last resort” if no connector discovering the specified device type matches the system.

Example: A connector with hdf.OnLastResort="Enclosure" will be activated if and only if no other connector matches AND has an Enclosure.Discovery.InstanceTable.