HWS-550
* Removed `logProductInformation` from the `main()` because when the
application's startup is a bit slower, the log4j2 context is refreshed
for some reason between logProductInformation and SpringBoot internal
logs and this led to two files generated (Minor problem but the fix is
always appreciated).

* logProductInformation is now called from the `ProductInfoService`
which is managed by Spring as the highest priority.

* Updated unit tests.

* Removed extra code and classes used to parse the internal
application.yml file. (We don't reinvent the wheel).